Deletar usuario e pessoa associada

+++ src/admin/users/UserManagementPage.tsx
@@ -81,8 +81,19 @@ const UserManagementPage = () => {
 
   const confirmDelete = async (userId: string) => {
     try {
-      console.log('Excluindo usuario: ', userId);
+      const userToDelete = users.find((user) => user.id === userId);
+      if (!userToDelete) {
+        Alert.alert('Erro', 'Usuário não encontrado.');
+        return;
+      }
+  
       await axios.delete(`${process.env.PUBLIC_API_URL}/usuario/${userId}`, {
+        headers: {
+          Authorization: `Bearer ${session?.token}`,
+        },
+      });
+  
+      await axios.delete(`${process.env.PUBLIC_API_URL}/pessoa/${userToDelete.pessoa.id}`, {
         headers: {
           Authorization: `Bearer ${session?.token}`,
         },
@@ -90,12 +101,13 @@ const UserManagementPage = () => {
 
       setUsers(users.filter((user) => user.id !== userId));
       Alert.alert('Sucesso', 'Usuário excluído com sucesso.');
-
+  
     } catch (err) {
-      console.error('Erro ao deletar postagem:', err.response.data);
-      setError('Erro ao deletar postagem. Verifique sua conexão e tente novamente.');
+      console.error('Erro ao deletar usuário', err.response?.data || err);
+      setError('Erro ao excluir. Verifique sua conexão e tente novamente.');
     }
   };
+  
 
   const renderItem = ({ item }: { item: User }) => (
     <View style={{ alignItems: 'center', marginHorizontal: 16 }}>
